Moved components folder into src folder
