+++ package.json
@@ -14,7 +14,7 @@
     "geolib": "^2.0.21",
     "react": "~15.3.2",
     "react-native": "github:exponentjs/react-native#sdk-10.1.2",
-    "react-native-blur": "^1.0.1",
-    "time-js": "0.0.5"
+    "react-redux": "^4.4.5",
+    "redux": "^3.6.0"
   }
 }
