+++ screens/RecipeListScreen.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RecipeList from '../components/RecipeList';
+import RecipeList from '../src/components/RecipeList';
 
 export default class RecipeListScreen extends React.Component {
   static route = {
