+++ screens/SignupScreen.js
@@ -6,7 +6,7 @@ import {
 import {
   ExponentLinksView,
 } from '@exponent/samples';
-import SignupForm from '../components/SignupForm'
+import SignupForm from '../src/components/SignupForm'
 
 export default class LinksScreen extends React.Component {
   static route = {
