+++ src/reducers/SessionReducer.js
@@ -1,0 +1,10 @@
+export default (state = false, action) => {
+  switch (action.type) {
+    case "login":
+      return action.payload;
+    case 'logout':
+      return action.payload;
+    default:
+      return state;
+  }
+};
