+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import SessionReducer from './SessionReducer';
+
+export default combineReducers({
+  currentUser: SessionReducer,
+});
